# Use native JSON parsing availible in Emacs 27 and report progress