Writing a huge amount of HTML files can take a while to finish. Inform
the user about the progress so the user can estimate remaining time
and clearly see that Emacs is busy working, not hung.

Speed Improvement
---------------------------
```elisp
;; Current implementation
(benchmark 1 '(devdocs-install '((name . "PHP") (slug . "php") (type . "php") (links (home . "https://www.php.net/") (code . "https://git.php.net/?p=php-src.git;a=summary")) (release . "8.1.5") (mtime . 1650899439) (db_size . 25993841))))
"Elapsed time: 17.163928s (6.059382s in 93 GCs)"
;; This commit
(benchmark 1 '(devdocs-install '((name . "PHP") (slug . "php") (type . "php") (links (home . "https://www.php.net/") (code . "https://git.php.net/?p=php-src.git;a=summary")) (release . "8.1.5") (mtime . 1650899439) (db_size . 25993841))))
"Elapsed time: 12.833806s (2.523597s in 39 GCs)"
```