@astoff 2022-05-18
Adding a progress reporter sounds good, thanks for the idea.

As to the JSON parsing, Emacs can be compiled without libjansson, in which case `json-parse-buffer` is not defined. Sadly, we'd have to check and define a fallback behavior manually if we want to use the native JSON parsing. I think this not worth the trouble in this particular application.