@astoff 2022-06-06
I've added a progress reporter in https://github.com/astoff/devdocs.el/commit/d5d0cfbfbcd037ef8f84f41b2adc3f5a23baa11f as you suggested here.

As is mentioned, the better JSON parser can be used only if a better compatibility layer is introduced somewhere else (maybe we can start using the compat library in the future).